Moving plots to a separate file #96
